modified service, resource, security constant, exception, filter etc.

+++ src/main/java/com/psl/supportportal/service/UserServiceImpl.java
@@ -1,9 +1,11 @@
 package com.psl.supportportal.service;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 
@@ -12,6 +14,7 @@ import javax.transaction.Transactional;
 
 import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.tomcat.util.http.fileupload.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,13 +28,15 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import static com.psl.supportportal.constant.FileConstant.*;
-
+import static org.springframework.http.MediaType.*;
 import static com.psl.supportportal.constant.UserImplConstant.*;
+import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
 import com.psl.supportportal.domain.User;
 import com.psl.supportportal.domain.UserPrincipal;
 import com.psl.supportportal.enumeration.Role;
 import com.psl.supportportal.exception.domain.EmailAlreadyExistException;
 import com.psl.supportportal.exception.domain.EmailNotFoundException;
+import com.psl.supportportal.exception.domain.NotAnImageFileException;
 import com.psl.supportportal.exception.domain.UserNotFoundException;
 import com.psl.supportportal.exception.domain.UsernameExistException;
 import com.psl.supportportal.repository.UserRepository;
@@ -176,7 +181,7 @@ public class UserServiceImpl implements UserService, UserDetailsService {
 
 	@Override
 	public User addNewUser(String firstName, String lastName, String username, String email, String role,
-			boolean isNotLocked, boolean isActive, MultipartFile profileImage) throws EmailAlreadyExistException, UsernameExistException, UserNotFoundException, IOException {
+			boolean isNotLocked, boolean isActive, MultipartFile profileImage) throws EmailAlreadyExistException, UsernameExistException, UserNotFoundException, IOException, NotAnImageFileException {
 		validateNewUserNameAndEmail(StringUtils.EMPTY, username, email);
 		User user = new User();
 		user.setUserId(generateUserId());
@@ -194,18 +199,23 @@ public class UserServiceImpl implements UserService, UserDetailsService {
 		user.setProfileImageUrl(getTemporaryProfileImageUrl(username));
 		userRepository.save(user);
 		saveProfileImage(user, profileImage);
+		logger.info("New user password : "+ password);
+		//emailService.sendPasswordEmail(firstName, password, email);
 		return user;
 	}
 
-	private void saveProfileImage(User user, MultipartFile profileImage) throws IOException {
+	private void saveProfileImage(User user, MultipartFile profileImage) throws IOException, NotAnImageFileException {
 		if(profileImage != null) {
+			if(!Arrays.asList(IMAGE_JPEG_VALUE, IMAGE_PNG_VALUE, IMAGE_GIF_VALUE).contains(profileImage.getContentType())) {
+				throw new NotAnImageFileException(profileImage.getOriginalFilename() + "is not an image file");
+			}
 			Path userFolder = Paths.get(USER_FOLDER + user.getUsername()).toAbsolutePath().normalize();
 			if(!Files.exists(userFolder)) {
 				Files.createDirectories(userFolder);
 				logger.info(DIRECTORY_CREATED + userFolder);
 			}
 			Files.deleteIfExists(Paths.get(USER_FOLDER + user.getUsername() +DOT+ JPG_EXTENSION));
-			Files.copy(profileImage.getInputStream() , userFolder.resolve(user.getUsername() +DOT+JPG_EXTENSION));
+			Files.copy(profileImage.getInputStream() , userFolder.resolve(user.getUsername() +DOT+JPG_EXTENSION), REPLACE_EXISTING);
 			user.setProfileImageUrl(setProfileImageUrl(user.getUsername()));
 			userRepository.save(user);
 			logger.info(FILE_SAVED_IN_FILE_SYSTEM + profileImage.getOriginalFilename());
@@ -224,7 +234,7 @@ public class UserServiceImpl implements UserService, UserDetailsService {
 
 	@Override
 	public User updateUser(String currentUsername, String newFirstName, String newLastName, String newUsername,
-			String newEmail, String role, boolean isNotLocked, boolean isActive, MultipartFile profileImage) throws EmailAlreadyExistException, UsernameExistException, UserNotFoundException, IOException {
+			String newEmail, String role, boolean isNotLocked, boolean isActive, MultipartFile profileImage) throws EmailAlreadyExistException, UsernameExistException, UserNotFoundException, IOException, NotAnImageFileException {
 		User currentUser = validateNewUserNameAndEmail(currentUsername, newUsername, newEmail);
 		currentUser.setUsername(newUsername);
 		currentUser.setActive(isActive);
@@ -240,8 +250,11 @@ public class UserServiceImpl implements UserService, UserDetailsService {
 	}
 
 	@Override
-	public void deleteUser(Long id) {
-		userRepository.deleteById(id);
+	public void deleteUser(String username) throws IOException {
+		User user = userRepository.findUserByUsername(username);
+		userRepository.deleteById(user.getId());
+		Path userFolder = Paths.get(USER_FOLDER + user.getUsername()).toAbsolutePath().normalize();
+		FileUtils.deleteDirectory(new File(userFolder.toString()));
 	}
 
 	@Override
@@ -253,11 +266,12 @@ public class UserServiceImpl implements UserService, UserDetailsService {
 		String password = generatePassword();
 		user.setPassword(encodePassword(password));
 		userRepository.save(user);
+		logger.info("New password : "+ password);
 		//emailService.sendPasswordEmail(user.getFirstName(), password, email);
 	}
 
 	@Override
-	public User updateProfileImage(String username, MultipartFile profileImage) throws EmailAlreadyExistException, UsernameExistException, UserNotFoundException, IOException {
+	public User updateProfileImage(String username, MultipartFile profileImage) throws EmailAlreadyExistException, UsernameExistException, UserNotFoundException, IOException, NotAnImageFileException {
 		User user = validateNewUserNameAndEmail(username, null, null);
 		saveProfileImage(user, profileImage);
 		return user;
